// src/main/resources/angular/src/app/api/model/userBO.ts
/**
 * MAAS API
 * This is the Mobility as a Service API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: avalls@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { RoleBO } from './roleBO';


export interface UserBO {
    boUserId?: number;
    fullName: string;
    email: string;
    isLdapUser: boolean;
    cities: any;
    city?: string;
    /**
     * Encrypted password
     */
    profile: any;
}